Add short country name to referendum data

Augments data with an additional column holding the common English name of the country in which the referendum took place.

Usage

add_country_name(data)

Arguments

data

RDB referendum data as returned by rfrnds(). A data frame that at minimum contains the column country_code (with ISO 3166-1 alpha-2 or ISO 3166-3 alpha-4 codes).

Value

A tibble.

See also

Examples

rdb::rfrnds(date_max = clock::date_today(zone = "UTC"),
            date_min = clock::date_today(zone = "UTC") |> clock::add_years(-1L),
            quiet = TRUE) |>
  rdb:::add_country_name() |>
  dplyr::select(id,
                starts_with("country_"))
#> # A tibble: 115 × 3
#>    id                       country_code country_name            
#>    <chr>                    <fct>        <fct>                   
#>  1 64c10be80b8bae0c78c7ec6a CF           Central African Republic
#>  2 64a3ca8b0b8bae0c78c7ebb8 ML           Mali                    
#>  3 64a2c7ee0b8bae0c78c7eba6 CH           Switzerland             
#>  4 64a2bcde0b8bae0c78c7eb81 CH           Switzerland             
#>  5 64a2ba730b8bae0c78c7eb7d CH           Switzerland             
#>  6 6458c01f0b8bae0c78c7ead2 CH           Switzerland             
#>  7 6458bdb10b8bae0c78c7eace CH           Switzerland             
#>  8 6458bcad0b8bae0c78c7eaca CH           Switzerland             
#>  9 6458ba8b0b8bae0c78c7eac4 CH           Switzerland             
#> 10 6458b95b0b8bae0c78c7eac0 CH           Switzerland             
#> # ℹ 105 more rows