Add is_former_country flag to referendum data

Augments data with an additional column is_former_country indicating whether or not the column country_code holds an ISO 3166-3 alpha-4 code referring to a historical country which ceased to exist. is_former_country being FALSE means country_code holds an ISO 3166-1 alpha-2 code instead.

Usage

add_former_country_flag(data)

Arguments

data

RDB referendum data as returned by rfrnds(). A data frame that at minimum contains the column country_code (with ISO 3166-1 alpha-2 or ISO 3166-3 alpha-4 codes).

Value

A tibble.

See also

Examples

rdb::rfrnds(country_code = c("CD", "ZRCD")) |>
  rdb:::add_former_country_flag() |>
  dplyr::select(any_of("id"),
                starts_with("country_"),
                is_former_country)
#> # A tibble: 6 × 4
#>   id                       country_code country_name is_former_country
#>   <chr>                    <fct>        <fct>        <lgl>            
#> 1 5bbbfe2f92a21351232e4955 CD           Congo        FALSE            
#> 2 5bbbed0992a21351232e14bd CD           Congo        FALSE            
#> 3 5bbbe90592a21351232e05fd CD           Congo        FALSE            
#> 4 5bbbe3a692a21351232df789 CD           Congo        FALSE            
#> 5 5bbbe2d292a21351232df407 CD           Congo        FALSE            
#> 6 5bbbe2cc92a21351232df263 CD           Congo        FALSE