Add period to referendum data

Augments data with an additional column holding the specified period in which the referendum took place. The new column is named after period and its values are always of type integer.

Usage

add_period(
  data,
  period = c("week", "month", "quarter", "year", "decade", "century")
)

Arguments

data

RDB referendum data as returned by rfrnds(). A data frame that at minimum contains the column date.

period

Type of period to add. One of "week", "month", "quarter", "year", "decade" or "century".

Value

A tibble.

Details

Note that the period types "year", "decade" and "century" are unique timespans, while "week", "month" and "quarter" are recurring timespans (e.g. every year has a week 1).

See also

Examples

rdb::rfrnds(quiet = TRUE) |>
  rdb::add_period() |>
  dplyr::select(id, date, week)
#> Error in rdb::rfrnds(quiet = TRUE): unused argument (quiet = TRUE)

rdb::rfrnds(quiet = TRUE) |>
  rdb::add_period("year") |>
  dplyr::select(id, date, year)
#> Error in rdb::rfrnds(quiet = TRUE): unused argument (quiet = TRUE)