Authenticate a user session for the RDB API

Creates a new user session token if necessary. The token is stored in the R option rdb.user_session_tokens, a tibble with the columns email, token and date_time_last_active.

Usage

auth_session(
  email = pal::pkg_config_val(key = "api_username", pkg = this_pkg),
  password = pal::pkg_config_val(key = "api_password", pkg = this_pkg),
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg),
  quiet = FALSE
)

Arguments

email

The e-mail address of the user for which a session should be created. A character scalar.

password

The password of the user for which a session should be created. A character scalar.

use_testing_server

Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.

quiet

Whether or not to suppress printing status output from internal processing.

Value

The user session token as a character scalar, invisibly.

Details

email and password default to the package configuration options api_username and api_password respectively.

User session tokens expire automatically after 15 days of inactivity.