Clean referendum attachments

Deletes obsolete attachment files from the specified S3-compatible object storage bucket. An attachment is considered obsolete if it isn't linked to any referendum in the RDB.

Usage

clean_rfrnd_attachments(
  s3_bucket = s3_bucket_attachments,
  s3_endpoint = s3_endpoint_url,
  s3_access_key = pal::pkg_config_val(key = "nocodb_s3_access_key", pkg = this_pkg),
  s3_access_secret = pal::pkg_config_val(key = "nocodb_s3_access_secret", pkg = this_pkg),
  run_dry = TRUE
)

Arguments

s3_bucket

Name of the S3-compatible object storage bucket where the referendum attachments reside.

s3_endpoint

S3-compatible endpoint URL where s3_bucket is located.

s3_access_key

S3-compatible access key ID with access to s3_bucket.

s3_access_secret

S3-compatible access secret with access to s3_bucket.

run_dry

Whether or not to display the files to be deleted on the console instead of actually deleting any file.

Value

A character vector of deleted file attachment URLs.

See also