Connect to RDB DBMS

Connects to RDB's database management system, a PostgreSQL server managed by neon.tech.

Usage

connect(
  dbname = pg_db,
  host = pal::pkg_config_val("pg_host"),
  user = pal::pkg_config_val("pg_user"),
  password = pal::pkg_config_val("pg_password"),
  sslmode = "verify-full"
)

Arguments

dbname

Name of the RDB PostgreSQL database. Defaults to "rdb".

host

RDB PostgreSQL (neon.tech) host address. Defaults to a read replica which has no write access to the database.

user

RDB PostgreSQL username for database operations performed via this package. To be able to edit the database, the user must have sufficient privileges. Defaults to the read-only user r_anon.

password

Password for pg_user. Defaults to the password for the read-only user r_anon.

sslmode

PostgreSQL connection parameter that determines whether or with what priority a secure TLS connection is negotiated with the server. See the official documentation for more information about the supported keywords.

Value

An S4 object that inherits from DBI::DBIConnection. This object is used to communicate with the database engine.

See also

Examples

con <- rdb::connect()

con |> DBI::dbListTables()
#>  [1] "actors"                       "administrative_units"         "countries"                    "languages"                    "legal_instruments"           
#>  [6] "legal_norms"                  "municipalities"               "nocodb_users"                 "options"                      "referendum_clusters"         
#> [11] "referendum_positions"         "referendum_questions"         "referendum_sub_votes"         "referendum_titles"            "referendum_types"            
#> [16] "referendum_types_legal_norms" "referendum_types_referendums" "referendum_urls"              "referendum_votes"             "referendums"                 
#> [21] "subnational_entities"         "supranational_entities"       "test"                         "test2"                        "topics"                      
#> [26] "topics_referendums"