Create NocoDB tables

Creates all the PostgreSQL database tables that must be first created once in NocoDB after the PostgreSQL database has been added as external data source (e.g. via reset_nocodb()) to let NocoDB register the necessary metadata for the created_by and updated_by columns (so they are properly auto-filled by NocoDB later on).

After running create_nocodb_tbls(), the created tables are supposed to be recreated outside of NocoDB via pg_init_main_tbls().

Usage

create_nocodb_tbls(hostname = nocodb_hostname)

Arguments

hostname

NocoDB server hostname. A character scalar.

Value

A tibble with metadata about the newly created tables.

See also