Delete referendums from the RDB

Deletes data from the RDB's referendums table.

Usage

delete_rfrnds(data, connection = connect(), disconnect = TRUE)

Arguments

data

Referendum data to be deleted. A tibble that at least must contain the column id.

connection

Database connection to a PostgreSQL DBMS as returned by connect(). An S4 object that inherits from DBI::DBIConnection.

disconnect

Whether or not to terminate the database connection when finished.

Value

The updated referendums table data as a tibble, invisibly.

See also