Get RDB data model

Get RDB data model

Usage

dm(
  connection = connect(user = "rdb_admin", password = pg_role_pw("rdb_admin")),
  incl_admin_tbls = TRUE,
  incl_aux_admin_tbls = FALSE,
  add_dm_meta = TRUE,
  learn_keys = TRUE
)

Arguments

connection

PqConnection
Database connection to a PostgreSQL DBMS as returned by connect().

incl_admin_tbls

Whether or not to include tables that are administered fully programmatically (administrative_units, supranational_entities, countries, subnational_entities, municipalities, languages, topics and nocodb_users).

incl_aux_admin_tbls

Whether or not to include the auxiliary administrative unit tables supranational_entities, countries, subnational_entities, municipalities and nocodb_users.

add_dm_meta

Whether or not to add colors and table descriptions to the returned data model object.

learn_keys

Set to TRUE to query the definition of primary and foreign keys from the database. Currently works only for Postgres/Redshift and SQL Server databases. The default attempts to query and issues an informative message.

Value

A data model object.

See also

Examples

# visualize the whole RDB model
rdb::dm() |> dm::dm_draw(view_type = "all")
#> Error in eval_select_indices(quo, table_names, unique = unique): Can't select tables that don't exist.
#> ✖ Table `electorate` doesn't exist.

# download all RDB tables into local object (e.g. for later processing)
rdb_dm <- rdb::dm() |> dplyr::collect()
#> Error in eval_select_indices(quo, table_names, unique = unique): Can't select tables that don't exist.
#> ✖ Table `electorate` doesn't exist.