Download file attachments

Downloads file attachments from the Referendum Database (RDB). ... the attachments list column returned by rfrnds().

Usage

download_attachments(
  data,
  path = ".",
  filename_from_title = FALSE,
  fileext_from_mimetype = !filename_from_title
)

Arguments

data

RDB referendum data as returned by rfrnds(). A data frame that at minimum contains the columns id and attachments.

path

Path to the directory where the downloaded file(s) are to be written to.

filename_from_title

Whether to use an attachment's title field value as the filename. Note that attachment titles are not guaranteed to be unique, i.e. there can be multiple file attachments with the same title (but differing content, of course). If FALSE, the referendum identifier plus a sequential number is used as an attachment's filename.

fileext_from_mimetype

Whether to add a file extension suffix to the filename which is determined from the attachment's mimetype field.

Value

data, invisibly.

See also

Examples

# get the first three referendums which contain attachments
d <-
  rdb::rfrnds() |>
  dplyr::filter(purrr::map_lgl(attachments, \(x) !is.null(x))) |>
  dplyr::filter(dplyr::row_number() %in% 1:3)

# download the attachments to the current directory
d |> rdb::download_attachments()

# and delete them again
d$attachments |>
  purrr::map(\(x) x$title) |>
  unlist() |>
  fs::file_delete()
#> Error in enc2utf8(path): argument is not a character vector