Download file attachment

Downloads a file attachment from the Referendum Database (RDB). The necessary s3_object_keys identifying individual files are found in the files list column returned by rfrnds().

Usage

download_file_attachment(
  s3_object_key,
  path = ".",
  use_original_filename = FALSE
)

Arguments

s3_object_key

Key uniquely identifying the file in the RDB Amazon S3 bucket. A character scalar.

path

Path where the downloaded file is written to.

use_original_filename

Whether to save the file attachment using its original filename as uploaded. Note that original filenames are not unique, i.e. there are multiple file attachments with the same original filename (but differing content, of course). If FALSE, s3_object_key is used as filename. Only relevant if path is a directory.

Value

A response object, invisibly.

See also

Examples

# get object keys
obj_keys <- ""

# download them to the current working dir
purrr::walk(obj_keys,
            rdb::download_file_attachment)
#> Error in map(.x, .f, ..., .progress = .progress): ℹ In index: 1.
#> Caused by error:
#> ! [ENOENT] Failed to copy '/tmp/RtmpWDXZ9f/file126f374d8aa4f5' to './file126f374d8aa4f5': no such file or directory

# and delete them again
fs::file_delete(obj_keys)
#> Error: [ENOENT] Failed to remove '': no such file or directory