Edit existing referendums in the RDB

Edits existing referendum entries in the API](https://github.com/zdaarau/c2d-app/blob/master/docs/services.md#3-referendum-routes) via its API.

Usage

edit_rfrnds(
  data,
  email = pal::pkg_config_val(key = "api_username", pkg = this_pkg),
  password = pal::pkg_config_val(key = "api_password", pkg = this_pkg),
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)

Arguments

data

Updated referendum data. A tibble that must contain an id column identifying the referendums to be edited plus any additional columns containing the new values to update the corresponding database fields with. Note that due to current API requirements, the following columns must always be supplied:

is_draft

electorate_total

electorate_abroad

votes_yes

votes_no

votes_empty

votes_invalid

email

The e-mail address of the RDB API user account to be used for authentication. A character scalar.

password

The password of the RDB API user account to be used for authentication. A character scalar.

use_testing_server

Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.

Value

data, invisibly.

Details

Note that adding/editing the column files is not supported, i.e. it is simply dropped from data.

See also