Count number of referendums per period

Counts the number of RDB referendums per desired period, optionally by additional columns specified via by.

Usage

n_rfrnds_per_period(
  data,
  by = NULL,
  period = c("week", "month", "quarter", "year", "decade", "century"),
  fill_gaps = TRUE,
  period_floor = NULL,
  period_ceiling = NULL,
  descending = FALSE
)

Arguments

data

RDB referendum data as returned by rfrnds(). A data frame that at minimum contains the column specified in period or the column date (to compute the period column), plus the one(s) specified via by (if any).

by

Optional data column(s) to group by before counting number of referendums. Tidy selections are supported.

period

Type of period to count referendums by. One of "week", "month", "quarter", "year", "decade" or "century".

fill_gaps

Whether or not to add zero-value rows to the result for period gaps in data.

period_floor

Lower period limit up to which gaps are filled. If NULL, the lower limit is set to the minimum of period present in data. Only relevant if fill_gaps = TRUE and period is set to a unique timespan type ("year", "decade" or "century").

period_ceiling

Upper period limit up to which gaps are filled. If NULL, the upper limit is set to the maximum of period present in data. Only relevant if fill_gaps = TRUE and period is set to a unique timespan type ("year", "decade" or "century").

descending

Whether to sort the resulting table by period in descending or in ascending order.

Value

A tibble.

Details

Note that the period types "year", "decade" and "century" are unique timespans, while "week", "month" and "quarter" are recurring timespans (e.g. every year has a week 1).

See also

Examples

rdb::rfrnds() |>
  rdb::n_rfrnds_per_period()
#> # A tibble: 0 × 2
#> # ℹ 2 variables: week <int>, n <int>

rdb::rfrnds() |>
  rdb::n_rfrnds_per_period(by = level)
#> # A tibble: 0 × 3
#> # ℹ 3 variables: level <chr>, week <int>, n <int>

# without filling gaps
rdb::rfrnds() |>
  rdb::n_rfrnds_per_period(by = level,
                           fill_gaps = FALSE)
#> # A tibble: 0 × 3
#> # ℹ 3 variables: level <chr>, week <int>, n <int>

# per decade and by multiple columns
rdb::rfrnds() |>
  rdb::n_rfrnds_per_period(by = c(level, type),
                           period = "decade")
#> # A tibble: 0 × 4
#> # ℹ 4 variables: level <chr>, type <int>, decade <int>, n <int>

# count ballot dates instead of referendums
rdb::rfrnds() |>
  rdb::as_ballot_dates() |>
  rdb::n_rfrnds_per_period()
#> # A tibble: 0 × 2
#> # ℹ 2 variables: week <int>, n <int>