Call NocoDB API

Returns the response from an API call to the RDB NocoDB server as a list.

Usage

ncdb_api(
  path,
  method = c("GET", "CONNECT", "DELETE", "HEAD", "OPTIONS", "PATCH", "POST", "PUT",
    "TRACE"),
  body_json = NULL,
  auto_unbox = TRUE,
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg),
  max_tries = 3L,
  verbosity = NULL
)

Arguments

path

NocoDB API endpoint path. A character scalar.

method

HTTP request method. One of "GET", "CONNECT", "DELETE", "HEAD", "OPTIONS", "PATCH", "POST", "PUT" or "TRACE".

body_json

Data to include as JSON in the HTTP request body. Set to NULL for an empty body.

auto_unbox

Whether or not to automatically "unbox" length-1 vectors in body_json to JSON scalars.

auth_token

NocoDB API authentication token. A character scalar.

max_tries

Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.

verbosity

How much information to print? This is a wrapper around req_verbose() that uses an integer to control verbosity:

0: no output

1: show headers

2: show headers and bodies

3: show headers, bodies, and curl status messages.

Use with_verbosity() to control the verbosity of requests that you can't affect directly.

Value

A list.

See also