Nest topics hierarchically

Takes a data frame with the three columns topic_tier_1, topic_tier_2 and topic_tier_3 and transforms it into a nested data frame with the two columns name and parent_name.

Usage

nest_topics(data)

Arguments

data

Political topics like data_topics.

Value

A tibble with the columns name and parent_name.

See also

Examples

rdb::data_topics |> rdb::nest_topics()
#> # A tibble: 175 × 2
#>    name                                parent_name     
#>    <chr>                               <chr>           
#>  1 constitution                        political system
#>  2 procedure for constitutional reform political system
#>  3 legislative procedure               political system
#>  4 electoral system                    political system
#>  5 government, administration          institutions    
#>  6 parliament                          institutions    
#>  7 courts                              institutions    
#>  8 national bank                       institutions    
#>  9 citizens' initiatives               citizens' rights
#> 10 referendum                          citizens' rights
#> # ℹ 165 more rows