Update PostgreSQL table

Update PostgreSQL table

Usage

pg_tbl_update(tbl, data, sweep = FALSE, tbl_name, connection = connect())

Arguments

tbl

Table to be updated. A data frame, tibble or tibble extension like tbl_lazy.

data

New or updated table data. A data frame that must contain at least the table's primary key column plus any additional columns with new values to update the corresponding database fields with.

sweep

Whether or not to also sweep the table, i.e. delete rows that are not contained in data.

tbl_name

Table name. One of "actors", "options", "legal_norms", "referendum_types", "referendums", "referendum_titles", "referendum_questions", "referendum_positions", "referendum_votes", "referendum_sub_votes", "countries", "subnational_entities", "municipalities" and "languages".

connection

Database connection to RDB's PostgreSQL DBMS. An S4 object that inherits from DBI::DBIConnection.

Value

The updated table data as a tibble, invisibly.