rdb package configuration metadata

A tibble with metadata of all possible rdb package configuration options. See pal::pkg_config_val() for more information.

Usage

pkg_config

Format

A tibble with the columns key, default_value, default_value_dynamic, require, and description.

Examples

rdb::pkg_config
#> # A tibble: 13 × 5
#>    key                          default_value default_value_dynamic require description                                                                         
#>    <chr>                        <list>        <chr>                 <lgl>   <glue>                                                                              
#>  1 pg_host                      <chr [1]>     NA                    TRUE    RDB PostgreSQL (neon.tech) host address. Defaults to a [read replica](https://neon.…
#>  2 pg_user                      <chr [1]>     NA                    TRUE    RDB PostgreSQL username for database operations performed via this package. To be a…
#>  3 pg_password                  <chr [1]>     NA                    TRUE    Password for `pg_user`. Defaults to the password for the read-only user `r_anon`.   
#>  4 pg_roles_csv_file            <NULL>        NA                    TRUE    File path to a UTF-8 encoded CSV file containing the columns `role` and `password` …
#>  5 neon_api_key                 <NULL>        NA                    TRUE    [API key](https://neon.tech/docs/manage/api-keys) for the RDB neon.tech project `ne…
#>  6 nocodb_host                  <chr [1]>     NA                    TRUE    RDB NocoDB server host address. Required for [rdb::reset_nocodb()].                 
#>  7 nocodb_email                 <NULL>        NA                    TRUE    E-mail address of the RDB NocoDB server's super admin user. Required for [rdb::rese…
#>  8 nocodb_password              <NULL>        NA                    TRUE    Password of the RDB NocoDB server's super admin user. Required for [rdb::reset_noco…
#>  9 nocodb_user_account_csv_file <NULL>        NA                    FALSE   File path to a UTF-8 encoded CSV file containing the columns `name`, `email` and `p…
#> 10 nocodb_s3_access_key         <NULL>        NA                    TRUE    ID of an access key granting read/write access to the S3-compatible object storage …
#> 11 nocodb_s3_access_secret      <NULL>        NA                    TRUE    Secret of an access key granting read/write access to the S3-compatible object stor…
#> 12 fly_io_api_token             <NULL>        NA                    TRUE    Fly.io [API token](https://fly.io/docs/flyctl/tokens/) granting access to the RDB o…
#> 13 global_max_cache_age         <chr [1]>     NA                    TRUE    Maximal timespan to preserve the package's [pkgpins](https://pkgpins.rpkg.dev/) cac…