Printify referendum data column names

Renames referendum data column names to be ready to print. Useful e.g. to create publication tables.

Usage

printify_col_names(data)

Arguments

data

RDB referendum data as returned by rfrnds().

Value

A tibble.

Details

Column names unknown to this function are not changed.

See also

Examples

rdb::rfrnds(country_code = "IT") |> rdb::printify_col_names()
#> # A tibble: 128 × 69
#>    `RDB ID`  Offic…¹ sudd.…² ISO 3…³ Country Subna…⁴ Munic…⁵ Polit…⁶ Date       Forme…⁷ Title Titel Titre Origi…⁸ Quest…⁹ Commi…˟ Result Appro…˟ Rejec…˟ Total…˟
#>    <chr>     <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord>   <date>     <lgl>   <chr> <chr> <chr> <chr>   <chr>   <chr>   <fct>    <dbl>   <dbl>   <int>
#>  1 63208eff… NA      it0620… IT      Italy   NA      NA      nation… 2022-06-12 FALSE   Judi… Just… NA    NA      NA      NA      no          NA      NA  5.09e7
#>  2 63208da3… NA      it0520… IT      Italy   NA      NA      nation… 2022-06-12 FALSE   Judi… Just… NA    NA      NA      NA      no          NA      NA  5.09e7
#>  3 63208be1… NA      it0420… IT      Italy   NA      NA      nation… 2022-06-12 FALSE   Judi… Just… NA    NA      NA      NA      no          NA      NA  5.09e7
#>  4 63208a15… NA      it0320… IT      Italy   NA      NA      nation… 2022-06-12 FALSE   Judi… Just… NA    NA      NA      NA      no          NA      NA  5.09e7
#>  5 632085cd… NA      it0220… IT      Italy   NA      NA      nation… 2022-06-12 FALSE   Judi… Just… NA    NA      NA      NA      no          NA      NA  5.09e7
#>  6 5f6c6d3f… NA      NA      IT      Italy   NA      NA      nation… 2020-09-21 FALSE   Down… NA    NA    NA      NA      NA      yes         NA      NA  5.10e7
#>  7 5d1cc245… NA      NA      IT      Italy   Lombar… NA      subnat… 2017-10-22 FALSE   Auto… NA    NA    NA      NA      NA      yes         NA      NA  7.90e6
#>  8 5d1cba4e… NA      NA      IT      Italy   Venetia NA      subnat… 2017-10-22 FALSE   Auto… NA    NA    NA      NA      NA      yes         NA      NA  4.07e6
#>  9 5d19ffd1… NA      NA      IT      Italy   NA      NA      nation… 2016-12-04 FALSE   Cons… NA    NA    NA      NA      NA      no          NA      NA  5.08e7
#> 10 5cb86b29… NA      NA      IT      Italy   NA      NA      nation… 2016-03-15 FALSE   Canc… NA    NA    NA      NA      NA      no          NA      NA  5.07e7
#> # … with 118 more rows, 49 more variables: `Electorate abroad` <int>, `Yes votes` <int>, `No votes` <int>, `Empty votes` <int>, `Invalid votes` <int>,
#> #   `Subterritorial votes` <list>, `Lower house yes votes` <int>, `Lower house no votes` <int>, `Lower house abstentions` <int>, `Upper house yes` <int>,
#> #   `Upper house No` <int>, `Upper house abstentions` <int>, `Government's position` <fct>, `Top-tier political topics` <list>,
#> #   `Second-tier political topics` <list>, `Third-tier political topics` <list>, Remarks <chr>, `File metadata` <list>, `sudd.ch URL` <chr>,
#> #   `Swissvotes URL` <chr>, Sources <chr>, `Draft?` <lgl>, `Creation date/time` <dttm>, `Updated date/time` <dttm>, Type <fct>, `Legal basis type` <ord>,
#> #   `Urgent legal basis?` <lgl>, `Binding?` <lgl>, `Counterproposal?` <lgl>, `Assembly?` <lgl>, `Trigger type` <fct>, `Trigger actor` <fct>,
#> #   `Tigger actor's political level` <ord>, `Trigger threshold` <chr>, `Trigger time limit` <chr>, `Object type` <fct>, `Object ` <fct>, …