Purge NocoDB

Stops the fly_app NocoDB instance, purges NocoDB's internal database (the Litestream replication stored under {s3_bucket}/nocodb) and then restarts the NocoDB instance, thereby resetting it to factory settings. Beware that this results in loss of all existing NocoDB-specific metadata (comments, audit logs etc.).

Usage

purge_nocodb(
  fly_app = "rdb-nocodb",
  s3_bucket = fly_app,
  s3_endpoint = s3_endpoint_url,
  s3_access_key = pal::pkg_config_val("nocodb_s3_access_key"),
  s3_access_secret = pal::pkg_config_val("nocodb_s3_access_secret"),
  ask = TRUE,
  quiet = FALSE
)

Arguments

fly_app

Fly.io app name.

s3_bucket

Name of the S3-compatible object storage bucket to which NocoDB's internal SQLite database is replicated to via Litestream.

s3_endpoint

S3-compatible endpoint URL where s3_bucket is located.

s3_access_key

S3-compatible access key ID with access to s3_bucket.

s3_access_secret

S3-compatible access secret with access to s3_bucket.

ask

Whether or not to ask for confirmation before proceeding. Only relevant if run interactively.

quiet

Whether or not to suppress printing status output from internal processing.

Value

fly_app, invisibly.

See also

Examples

if (FALSE) { # \dontrun{
# purge NocoDB production instance
rdb:::purge_nocodb()

# purge NocoDB testing instance
rdb:::purge_nocodb(fly_app = "rdb-nocodb-testing")} # }