Read SQL script as statements

Reads in an SQL script and returns it as individual interpolated SQL statements using split_sql_str().

Usage

read_sql_file(.path, ..., .connection = connect())

Arguments

.path

Path to a file containing SQL statements.

...

Further (named) arguments used as temporary variables available for substitution while interpolating the SQL statements via DBI::sqlInterpolate().

.connection

Database connection as returned by connect(). An S4 object that inherits from DBI::DBIConnection.

Value

A list of DBI::SQL objects.

See also