Read in RDB table

Fetches a table from the RDB DBMS's public schema and returns it as a tibble.

Usage

read_tbl(
  tbl_name = tbl_metadata$name,
  connection = connect(),
  disconnect = TRUE
)

Arguments

tbl_name

Table name. One of "actors", "options", "legal_norms", "referendum_types", "referendums", "referendum_titles", "referendum_questions", "referendum_positions", "referendum_votes", "referendum_sub_votes", "countries", "subnational_entities", "municipalities" and "languages".

connection

Database connection to RDB's PostgreSQL DBMS. An S4 object that inherits from DBI::DBIConnection.

disconnect

Whether or not to terminate the database connection when finished.

Value

A tibble.

See also

Examples

rdb::read_tbl("countries")
#> # A tibble: 250 × 7
#>    code  code_long code_numeric name             name_long                       created_at          updated_at         
#>    <chr> <chr>     <chr>        <chr>            <chr>                           <dttm>              <dttm>             
#>  1 BD    BGD       050          Bangladesh       People's Republic of Bangladesh 2024-05-07 11:35:27 2024-05-07 15:07:35
#>  2 BE    BEL       056          Belgium          Kingdom of Belgium              2024-05-07 11:35:27 2024-05-07 15:07:35
#>  3 BF    BFA       854          Burkina Faso     Burkina Faso                    2024-05-07 11:35:27 2024-05-07 15:07:35
#>  4 BG    BGR       100          Bulgaria         Republic of Bulgaria            2024-05-07 11:35:27 2024-05-07 15:07:35
#>  5 BH    BHR       048          Bahrain          Kingdom of Bahrain              2024-05-07 11:35:27 2024-05-07 15:07:35
#>  6 BI    BDI       108          Burundi          Republic of Burundi             2024-05-07 11:35:27 2024-05-07 15:07:35
#>  7 BJ    BEN       204          Benin            Republic of Benin               2024-05-07 11:35:27 2024-05-07 15:07:35
#>  8 BL    BLM       652          Saint Barthélemy Saint Barthélemy                2024-05-07 11:35:27 2024-05-07 15:07:35
#>  9 BM    BMU       060          Bermuda          Bermuda                         2024-05-07 11:35:27 2024-05-07 15:07:35
#> 10 BN    BRN       096          Brunei           Brunei Darussalam               2024-05-07 11:35:27 2024-05-07 15:07:35
#> # ℹ 240 more rows