Reset NocoDB

Completely re-creates the connection settings and Main base and adds the RDB team's user accounts on the specified NocoDB server. Beware that this results in loss of all existing NocoDB-specific metadata (comments, audit logs etc.).

Usage

reset_nocodb(
  hostname = pal::pkg_config_val(key = "nocodb_host", pkg = this_pkg),
  ask = TRUE,
  save_user_mapping = TRUE,
  quiet = TRUE
)

Arguments

hostname

NocoDB server hostname. A character scalar.

ask

Whether or not to ask for confirmation before deleting the existing RDB NocoDB base. Only relevant if run interactively.

save_user_mapping

Whether or not to add NocoDB's current user data to the nocodb_users table via update_nocodb_users() before resetting it.

quiet

Whether or not to suppress printing status output from internal processing.

Value

hostname, invisibly.

See also