Reset RDB

Resets the RDB NocoDB server, optionally including the whole RDB PostgreSQL database and/or tables.

Usage

reset_rdb(
  hostname_nocodb = pal::pkg_config_val("nocodb_host"),
  hostname_pg = pal::pkg_config_val("pg_host"),
  reset_db = FALSE,
  reset_tbls = reset_db,
  ask = TRUE,
  disconnect = TRUE,
  quiet = FALSE
)

Arguments

hostname_nocodb

NocoDB server hostname. A character scalar.

hostname_pg

PostgreSQL server hostname. A character scalar.

reset_db

Whether or not to first fully recreate the RDB PostgreSQL database using pg_reset_db() and pg_init_db(). If FALSE, only the RDB PostgreSQL tables are recreated.

reset_tbls

Whether or not to recreate the RDB PostgreSQL tables.

ask

Whether or not to ask for confirmation before proceeding. Only relevant if run interactively.

disconnect

logical(1)
Whether or not to terminate the database connection when finished.

quiet

Whether or not to suppress printing status output from internal processing.

Value

hostname_nocodb, invisibly.

See also

Examples

if (FALSE) { # \dontrun{
# reset production PGSQL DB branch and NocoDB instance
rdb::reset_rdb(reset_db = TRUE)

# reset testing PGSQL DB branch and NocoDB instance
rdb::reset_rdb(hostname_nocodb = Sys.getenv("R_RDB_NOCODB_HOST_TESTING"),
               hostname_pg = Sys.getenv("R_RDB_PG_HOST_TESTING"),
               reset_db = TRUE)} # }