Get legal bases data

Downloads the legal bases data from the Referendum Database (RDB). See the codebook for a detailed description of all variables.

Usage

rfrnd_types(
  connection = connect(),
  disconnect = TRUE,
  use_cache = TRUE,
  max_cache_age = "1 week"
)

Arguments

connection

Database connection to a PostgreSQL DBMS as returned by connect(). An S4 object that inherits from DBI::DBIConnection.

disconnect

Whether or not to terminate the database connection when finished.

use_cache

Whether or not to return cached results if possible. If FALSE, results are always newly fetched regardless of max_cache_age.

max_cache_age

Duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Use Inf to disable cache expiry. Only relevant if use_cache = TRUE.

Value

A tibble.

See also

Examples

# get all legal bases
rdb::rfrnd_types()
#> # A tibble: 13 × 12
#>       id country_code level    subnational_entity_code municipality_id title valid_from valid_to created_by updated_by created_at          updated_at         
#>    <int> <chr>        <chr>    <chr>                   <chr>           <chr> <date>     <date>   <chr>      <chr>      <dttm>              <dttm>             
#>  1     2 CH           national NA                      NA              mand… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#>  2     3 CH           national NA                      NA              mand… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#>  3     4 CH           national NA                      NA              mand… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#>  4     5 CH           national NA                      NA              mand… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#>  5     6 CH           national NA                      NA              mand… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#>  6     7 CH           national NA                      NA              mand… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#>  7     8 CH           national NA                      NA              opti… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#>  8     9 CH           national NA                      NA              opti… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#>  9    10 CH           national NA                      NA              opti… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#> 10    11 CH           national NA                      NA              opti… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#> 11    12 CH           national NA                      NA              opti… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#> 12    13 CH           national NA                      NA              opti… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-13 01:57:47
#> 13     1 CH           national NA                      NA              mand… 1-01-01    NA       r_user     r_user     2024-06-13 01:57:47 2024-06-14 18:05:05