Generate RDB languages table

Updates the RDB's languages table with the data from data_iso_639_1.

Usage

update_langs(sweep = TRUE, connection = connect(), disconnect = TRUE)

Arguments

sweep

Whether or not to also sweep the table, i.e. delete rows that are not contained in data.

connection

Database connection to RDB's PostgreSQL DBMS. An S4 object that inherits from DBI::DBIConnection.

disconnect

Whether or not to terminate the database connection when finished.

Value

The updated languages table data as a tibble, invisibly.

See also