Update RDB referendums table

Updates the RDB's referendums table with data. By default, rows not contained in data are kept.

Usage

update_rfrnds(data, sweep = FALSE, connection = connect(), disconnect = TRUE)

Arguments

data

New or updated referendum data. A tibble that must contain a id column identifying the referendums to be updated plus any additional columns containing the new values to update the corresponding database fields with.

sweep

Whether or not to also sweep the table, i.e. delete rows that are not contained in data.

connection

Database connection to a PostgreSQL DBMS as returned by connect(). An S4 object that inherits from DBI::DBIConnection.

disconnect

Whether or not to terminate the database connection when finished.

Value

data, invisibly.

See also