Upload attachments to NocoDB

Uploads one or more files to the RDB NocoDB server as attachments via the POST /api/v2/storage/upload API endpoint. Uploaded files are stored under /nc/uploads/ in the Backblaze B2 bucket rdb-attachments.

Usage

upload_attachments_to_ncdb(
  paths,
  names = fs::path_file(paths),
  types = mime::guess_type(paths),
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg),
  max_tries = 3L,
  verbosity = NULL
)

Arguments

paths

Path(s) to the file(s) to be uploaded. A character vector.

names

File name(s) to assign to the uploaded file(s). A character vector. Note that a random string like _NnW3C is always appended to the resulting file name before its file type suffix, so name = "some-pic.jpg" will result in something like some-pic_NnW3C.jpg, for example.

types

Media type(s) of the uploaded files. A character vector.

auth_token

NocoDB API authentication token. A character scalar.

max_tries

Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.

verbosity

How much information to print? This is a wrapper around req_verbose() that uses an integer to control verbosity:

0: no output

1: show headers

2: show headers and bodies

3: show headers, bodies, and curl status messages.

Use with_verbosity() to control the verbosity of requests that you can't affect directly.

Value

A tibble containing metadata about the uploaded attachments.

See also