Get unnested variable names

Returns the unnested analogue(s) of the specified variable name(s), which result from unnesting. For variable names that do not refer to nested list columns, var_names is simply returned as-is.

Usage

var_name_unnested(var_names)

Arguments

var_names

Variable name(s). Those not present in data_codebook remain untouched. A character vector.

Value

A character vector of the same length as var_names.

See also

Examples

rdb::var_name_unnested("inst_object_revision_modes")
#> [1] "inst_object_revision_mode"
rdb::var_name_unnested(paste0("topics_tier_", 1:3))
#> [1] "topic_tier_1" "topic_tier_2" "topic_tier_3"