Add turnout to referendum data

Augments data with an additional column turnout containing the voter turnout calculated as:

votes_yes+votes_no+votes_empty+votes_invalidelectorate_total\frac{votes\_yes+votes\_no+votes\_empty+votes\_invalid}{electorate\_total}electorate_totalvotes_yes+votes_no+votes_empty+votes_invalid​

Usage

add_turnout(data, rough = TRUE, excl_dubious = TRUE)

Arguments

data

RDB referendum data as returned by rfrnds(). A data frame that at minimum contains the columns electorate_total, votes_yes, votes_no, votes_empty and votes_invalid.

rough

Whether to fall back on a "rough" calculation of the turnout in case any of the variables votes_empty or votes_invalid is unknown (NA), or to be strict and return NA in such a case.

excl_dubious

Whether or not to exclude obviously dubious turnout numbers (those > 1.0) by setting them to NA. Such numbers stem either from data errors or (officially) tampered numbers.

Value

A tibble.

See also

Examples

# rough turnout numbers
rdb::rfrnds(country_code = "IT",
            quiet = TRUE) |>
  rdb::add_turnout() |>
  dplyr::select(id,
                electorate_total,
                starts_with("votes_"),
                turnout)
#> # A tibble: 128 × 8
#>    id                       electorate_total votes_yes votes_no votes_empty votes_invalid votes_per_subterritory turnout
#>    <chr>                               <int>     <int>    <int>       <int>         <int> <list>                   <dbl>
#>  1 63208effa52c3995043a9101         50910473   6776491  2684184      721157        239995 <NULL>                   0.205
#>  2 63208da3a52c3995043a90fb         50910473   6715104  2705341      752871        246373 <NULL>                   0.205
#>  3 63208be1a52c3995043a90f6         50910473   6968829  2543081      682995        232129 <NULL>                   0.205
#>  4 63208a15a52c3995043a90eb         50910473   5271228  4249305      677146        226959 <NULL>                   0.205
#>  5 632085cda52c3995043a90e3         50910473   5093092  4501877      616268        220564 <NULL>                   0.205
#>  6 5f6c6d3fd1291cc3961f1b8b         50955985  17913259  7691837      218093        226569 <NULL>                   0.511
#>  7 5d1cc245dc0aeb467bfcbfb6          7897056   2882531   119420          NA         23151 <NULL>                   0.383
#>  8 5d1cba4edc0aeb467bfcbfad          4068560   2273985    43938        5165          5865 <NULL>                   0.572
#>  9 5d19ffd1fe65c2bfeba00bb9         50773284  13431087 19421025       83418        308728 <NULL>                   0.655
#> 10 5cb86b29cb48652399618eda         50681765  13334607  2198715      104413        168753 <NULL>                   0.312
#> # ℹ 118 more rows

# strict turnout numbers
rdb::rfrnds(country_code = "IT",
            quiet = TRUE) |>
  rdb::add_turnout(rough = FALSE) |>
  dplyr::select(id,
                electorate_total,
                starts_with("votes_"),
                turnout)
#> # A tibble: 128 × 8
#>    id                       electorate_total votes_yes votes_no votes_empty votes_invalid votes_per_subterritory turnout
#>    <chr>                               <int>     <int>    <int>       <int>         <int> <list>                   <dbl>
#>  1 63208effa52c3995043a9101         50910473   6776491  2684184      721157        239995 <NULL>                   0.205
#>  2 63208da3a52c3995043a90fb         50910473   6715104  2705341      752871        246373 <NULL>                   0.205
#>  3 63208be1a52c3995043a90f6         50910473   6968829  2543081      682995        232129 <NULL>                   0.205
#>  4 63208a15a52c3995043a90eb         50910473   5271228  4249305      677146        226959 <NULL>                   0.205
#>  5 632085cda52c3995043a90e3         50910473   5093092  4501877      616268        220564 <NULL>                   0.205
#>  6 5f6c6d3fd1291cc3961f1b8b         50955985  17913259  7691837      218093        226569 <NULL>                   0.511
#>  7 5d1cc245dc0aeb467bfcbfb6          7897056   2882531   119420          NA         23151 <NULL>                  NA    
#>  8 5d1cba4edc0aeb467bfcbfad          4068560   2273985    43938        5165          5865 <NULL>                   0.572
#>  9 5d19ffd1fe65c2bfeba00bb9         50773284  13431087 19421025       83418        308728 <NULL>                   0.655
#> 10 5cb86b29cb48652399618eda         50681765  13334607  2198715      104413        168753 <NULL>                   0.312
#> # ℹ 118 more rows