Transform to ballot-date-level observations

Transforms referendum-level observations to ones on the level of ballot date and jurisdiction via nesting of referendum-level columns. The individual values of all the referendums on a specific ballot date in a specific jurisdiction are preserved in a list column named rfrnd_data.

Usage

as_ballot_dates(data)

Arguments

data

RDB referendum data as returned by rfrnds(). A data frame that at minimum contains the column date.

Value

A tibble.

See also

Examples

rdb::rfrnds(country_code = "AT",
            quiet = TRUE) |>
  rdb::as_ballot_dates()
#> # A tibble: 9 × 8
#>   country_code country_name subnational_entity_name  municipality level       date       is_former_country rfrnd_data       
#>   <fct>        <fct>        <chr>                    <chr>        <ord>       <date>     <lgl>             <list>           
#> 1 AT           Austria      NA                       NA           national    2013-01-20 FALSE             <tibble [2 × 62]>
#> 2 AT           Austria      NA                       NA           national    1994-06-12 FALSE             <tibble [1 × 62]>
#> 3 AT           Austria      NA                       NA           national    1978-11-05 FALSE             <tibble [1 × 62]>
#> 4 AT           Austria      NA                       NA           national    1938-04-10 FALSE             <tibble [1 × 62]>
#> 5 AT           Austria      Salzburg                 NA           subnational 1921-05-29 FALSE             <tibble [1 × 62]>
#> 6 AT           Austria      Tyrol                    NA           subnational 1921-04-24 FALSE             <tibble [1 × 62]>
#> 7 AT           Austria      Carinthia                NA           subnational 1920-10-10 FALSE             <tibble [2 × 62]>
#> 8 AT           Austria      Klagenfurt Becken Zone I NA           subnational 1920-10-10 FALSE             <tibble [1 × 62]>
#> 9 AT           Austria      Vorarlberg               NA           subnational 1919-05-11 FALSE             <tibble [1 × 62]>