Connect to RDB DBMS

Connects to RDB's database management system, a PostgreSQL server managed by neon.tech.

Usage

connect(
  host = pal::pkg_config_val(key = "pg_host", pkg = this_pkg),
  user = pal::pkg_config_val(key = "pg_user", pkg = this_pkg),
  password = pal::pkg_config_val(key = "pg_password", pkg = this_pkg),
  sslmode = "verify-full"
)

Arguments

host

RDB PostgreSQL (neon.tech) host address. Defaults to a read replica which has no write access to the database.

user

RDB PostgreSQL username. To be able to edit the database, the user must have sufficient privileges. Defaults to the read-only user r_anon.

password

Password for the specified RDB PostgreSQL user. Defaults to the password for user r_anon.

sslmode

PostgreSQL connection parameter that determines whether or with what priority a secure TLS connection is negotiated with the server. See the official documentation for more information about the supported keywords.

Value

An S4 object that inherits from DBI::DBIConnection. This object is used to communicate with the database engine.

See also

Examples

con <- rdb::connect()

con |> DBI::dbListTables()
#> [1] "countries"            "languages"            "municipalities"       "subnational_entities"