Get RDB data model

Get RDB data model

Usage

dm(
  connection = connect(user = "rdb_admin", password = pg_role_pw("rdb_admin")),
  incl_admin_tbls = TRUE,
  incl_aux_admin_tbls = FALSE
)

Arguments

connection

PqConnection
Database connection to a PostgreSQL DBMS as returned by connect().

incl_admin_tbls

Whether or not to include tables that are administered fully programmatically (administrative_units, supranational_entities, countries, subnational_entities, municipalities, languages, topics and nocodb_users).

incl_aux_admin_tbls

Whether or not to include the auxiliary administrative unit tables supranational_entities, countries, subnational_entities, municipalities and nocodb_users.

Value

A data model object.

See also

Examples

# visualize the whole RDB model
rdb::dm() |> dm::dm_draw(view_type = "all")

# download all RDB tables into local object (e.g. for later processing)
rdb_dm <- rdb::dm() |> dplyr::collect()