Hierarchize topics

Reconstructs the hierarchical relations between the three topic variables topics_tier_1, topics_tier_2 and topics_tier_3. Can also be used to simply determine the parent topic(s) of any topic.

Usage

hierarchize_topics(x)

Arguments

x

The topics to hierarchize. Either a character vector of topics or a single-row data frame containing at least the columns topics_tier_1, topics_tier_2 and topics_tier_3.

Value

A tibble with the columns topic_tier_1, topic_tier_2, and topic_tier_3.

See also

Examples

rdb::hierarchize_topics("territorial questions")
#>         topic_tier_1 topic_tier_2          topic_tier_3
#> 1 state organisation   federalism territorial questions

# hierarchize the topics of all Austrian referendums
rdb::rfrnds() |>
  dplyr::filter(country_code == "AT") |>
  dplyr::group_split(id) |>
  purrr::map(rdb::hierarchize_topics)
#> Error in dplyr::filter(rdb::rfrnds(), country_code == "AT"): ℹ In argument: `country_code == "AT"`.
#> Caused by error:
#> ! object 'country_code' not found