List referendum territories from sudd.ch

Lists all referendum territories from sudd.ch, which means each country_name_de together with all the associated territory_name_de, their search URL and their number of occurrences.

Usage

list_sudd_territories()

Value

A tibble.

Details

Note that the values in the territory_name_de column returned by this function can differ from those in the territory_name_de column of sudd_rfrnds() and list_sudd_rfrnds(). The latter is often more extensive and usually includes the country_name_de (in parentheses) for subnational referendums.

About sudd.ch

sudd stands for Suchmaschine für direkte Demokratie (German) and is operated by Beat Müller. Its database content is licensed under Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0).

See also

Examples

rdb::list_sudd_territories()
#> # A tibble: 431 × 4
#>    country_name_de     territory_name_de    filter_url                                                    n    
#>    <chr>               <chr>                <chr>                                                         <chr>
#>  1 Albanien            Albanien             https://sudd.ch/list.php?de=de&area=albanien                  3    
#>  2 Algerien            Algerien             https://sudd.ch/list.php?de=de&area=algerien                  13   
#>  3 Andorra             Andorra              https://sudd.ch/list.php?de=de&area=andorra                   5    
#>  4 Anguilla            Anguilla             https://sudd.ch/list.php?de=de&area=anguilla                  2    
#>  5 Antigua und Barbuda Antigua und Barbuda  https://sudd.ch/list.php?de=de&area=antigua%20und%20barbuda   1    
#>  6 Argentinien         Argentinien          https://sudd.ch/list.php?de=de&area=argentinien               2    
#>  7 Argentinien         Provinz Buenos Aires https://sudd.ch/list.php?de=de&area=provinz%20buenos%20aires  1    
#>  8 Argentinien         Provinz Corrientes   https://sudd.ch/list.php?de=de&area=provinz%20corrientes      1    
#>  9 Argentinien         Valle 16 de Octubre  https://sudd.ch/list.php?de=de&area=valle%2016%20de%20octubre 1    
#> 10 Armenien            Armenien             https://sudd.ch/list.php?de=de&area=armenien                  5    
#> # ℹ 421 more rows