Update PostgreSQL table

Updates the specified PostgreSQL tbl with values from data.

Usage

pg_tbl_update(
  tbl,
  data,
  tbl_name,
  schema = pg_schema,
  sweep = FALSE,
  override_pk_sys_val = FALSE,
  connection = connect()
)

Arguments

tbl

PostgreSQL database table to be updated. A dbplyr::tbl.src_dbi object.

data

New or updated table data. A data frame that must contain at least the table's primary key column plus any additional columns with new values to update the corresponding database fields with.

tbl_name

character(1)
Table name. One of

"actors"

"options"

"legal_instruments"

"legal_norms"

"referendum_types"

"referendum_clusters"

"referendums"

"referendum_types_legal_norms"

"referendum_types_referendums"

"topics_referendums"

"referendum_titles"

"referendum_questions"

"referendum_urls"

"referendum_positions"

"electorate"

"referendum_votes"

"referendum_sub_votes"

"administrative_units"

"supranational_entities"

"countries"

"subnational_entities"

"municipalities"

"languages"

"topics"

"nocodb_users"

sweep

Whether or not to also sweep the table, i.e. delete rows that are not contained in data.

override_pk_sys_val

Whether or not to add new rows that have primary key values set (i.e. which are not NA) to a tbl whose primary key is GENERATED ALWAYS. Setting override_pk_sys_val = TRUE is roughly equivalent to specifying OVERRIDING SYSTEM VALUE in an SQL INSERT statement.

connection

PqConnection
Database connection to a PostgreSQL DBMS as returned by connect().

Value

The updated table data as a tibble, invisibly. Note that newly inserted rows aren't included if the table's primary key column is GENERATED ALWAYS.

Details

If the tbl's primary key column is GENERATED ALWAYS, rows to be newly added must have NA values in the corresponding data column unless override_pk_sys_val = TRUE. This is a safety measure to avoid accidental insertions.

See also