rdb: Download Data from the Referendum Database (RDB), Which Covers Direct Democratic Votes Worldwide

rdb provides access to the Referendum Database (RDB) from R. This database aims to record all direct democratic votes worldwide and is operated by the Centre for Democracy Studies Aarau (ZDA) at the University of Zurich, Switzerland.

Package configuration

Some of rdb's functionality is controlled via package-specific global configuration which can either be set via R options or environment variables (the former take precedence). This configuration includes:

Description	R option	Environment variable	Default value

RDB Services API username	rdb.api_username	RDB_API_USERNAME

RDB Services API password	rdb.api_password	RDB_API_PASSWORD

Maximal timespan to preserve the package's pkgpins cache. Cache entries older than this will be deleted upon package loading.	rdb.global_max_cache_age	RDB_GLOBAL_MAX_CACHE_AGE	"30 days"

Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.	rdb.use_testing_server	RDB_USE_TESTING_SERVER	FALSE

Whether or not to run the tests that use the testing servers for RDB Services API calls etc. during devtools::test().	rdb.test_testing_server	RDB_TEST_TESTING_SERVER	FALSE

See also

Author

Maintainer: Salim Brüggemann salim-b@pm.me (ORCID)

Authors:

Uwe Serdült uwe.serduelt@zda.uzh.ch (ORCID)

Other contributors:

Centre for Democracy Studies Aarau (ZDA) info@zdaarau.ch [copyright holder]