rdb: Download Data from the Referendum Database (RDB), Which Covers Direct Democratic Votes Worldwide

rdb provides access to the Referendum Database (RDB) from R. This database aims to record all direct democratic votes worldwide and is operated by the Centre for Democracy Studies Aarau (ZDA) at the University of Zurich, Switzerland.

Package configuration

Some of rdb's functionality is controlled via package-specific global configuration which can either be set via R options or environment variables (the former take precedence). This configuration includes:

Description	R option	Environment variable	Default value

RDB PostgreSQL (neon.tech) host address. Defaults to a read replica which has no write access to the database.	rdb.pg_host	R_RDB_PG_HOST	"ep-frosty-flower-a28keh10.eu-central-1.aws.neon.tech"

RDB PostgreSQL username for database operations performed via this package. To be able to edit the database, the user must have sufficient privileges. Defaults to the read-only user r_anon.	rdb.pg_user	R_RDB_PG_USER	"r_anon"

Password for pg_user. Defaults to the password for the read-only user r_anon.	rdb.pg_password	R_RDB_PG_PASSWORD	"WTYUeDtTm0UV4nXzTZ7morHSdM0wQh0p"

File path to a UTF-8 encoded CSV file containing the columns role and password describing additional roles on the RDB PostgreSQL server. Required for pg_init_db() and others.	rdb.pg_roles_csv_file	R_RDB_PG_ROLES_CSV_FILE

API key for the RDB neon.tech project neon_rdb. Required for restart_neon_ep().	rdb.neon_api_key	R_RDB_NEON_API_KEY

RDB NocoDB server host address. Required for reset_nocodb().	rdb.nocodb_host	R_RDB_NOCODB_HOST	"admin.rdb.vote"

E-mail address of the RDB NocoDB server's super admin user. Required for reset_nocodb().	rdb.nocodb_email	R_RDB_NOCODB_EMAIL

Password of the RDB NocoDB server's super admin user. Required for reset_nocodb().	rdb.nocodb_password	R_RDB_NOCODB_PASSWORD

File path to a UTF-8 encoded CSV file containing the columns name, email and password describing additional user accounts to create on the RDB NocoDB server via reset_nocodb().	rdb.nocodb_user_account_csv_file	R_RDB_NOCODB_USER_ACCOUNT_CSV_FILE

ID of an access key granting read/write access to the S3-compatible object storage buckets rdb-attachments and rdb-nocodb. Required for various administration functions.	rdb.nocodb_s3_access_key	R_RDB_NOCODB_S3_ACCESS_KEY

Secret of an access key granting read/write access to the S3-compatible object storage buckets rdb-attachments and rdb-nocodb. Required for various administration functions.	rdb.nocodb_s3_access_secret	R_RDB_NOCODB_S3_ACCESS_SECRET

Fly.io API token granting access to the RDB organization. Required for restart_postgrest().	rdb.fly_io_api_token	R_RDB_FLY_IO_API_TOKEN

Maximal timespan to preserve the package's pkgpins cache. Cache entries older than this will be deleted upon package loading.	rdb.global_max_cache_age	R_RDB_GLOBAL_MAX_CACHE_AGE	"30 days"

See also

Author

Maintainer: Salim Brüggemann salim-b@pm.me (ORCID)

Authors:

Uwe Serdült uwe.serduelt@zda.uzh.ch (ORCID)

Other contributors:

Centre for Democracy Studies Aarau (ZDA) info@zdaarau.ch [copyright holder]