rdb: Download Data from the Referendum Database (RDB), Which Covers Direct Democratic Votes Worldwide

rdb provides access to the Referendum Database (RDB) from R. This database aims to record all direct democratic votes worldwide and is operated by the Centre for Democracy Studies Aarau (ZDA) at the University of Zurich, Switzerland.

Package configuration

Some of rdb's functionality is controlled via package-specific global configuration which can either be set via R options or environment variables (the former take precedence). This configuration includes:

Description	R option	Environment variable	Default value

Hostname of the RDB NocoDB server. Used as default for the *nococdb*() family of functions.	rdb.nocodb_hostname	R_RDB_NOCODB_HOSTNAME

NocoDB API authentication token for the RDB NocoDB server. Used as default for the *nococdb*() family of functions.	rdb.nocodb_api_token	R_RDB_NOCODB_API_TOKEN

Title (name) of the main base on the RDB NocoDB server. Used as default for the *nococdb*() family of functions.	rdb.nocodb_base_title	R_RDB_NOCODB_BASE_TITLE	"Main"

RDB PostgreSQL (neon.tech) host address. Defaults to a read replica which has no write access to the database.	rdb.pg_host	R_RDB_PG_HOST	"ep-frosty-flower-a28keh10.eu-central-1.aws.neon.tech"

RDB PostgreSQL username. To be able to edit the database, the user must have sufficient privileges. Defaults to the read-only user r_anon.	rdb.pg_user	R_RDB_PG_USER	"r_anon"

Password for the specified RDB PostgreSQL user. Defaults to the password for user r_anon.	rdb.pg_password	R_RDB_PG_PASSWORD	"WTYUeDtTm0UV4nXzTZ7morHSdM0wQh0p"

Maximal timespan to preserve the package's pkgpins cache. Cache entries older than this will be deleted upon package loading.	rdb.global_max_cache_age	R_RDB_GLOBAL_MAX_CACHE_AGE	"30 days"

See also

Author

Maintainer: Salim Brüggemann salim-b@pm.me (ORCID)

Authors:

Uwe Serdült uwe.serduelt@zda.uzh.ch (ORCID)

Other contributors:

Centre for Democracy Studies Aarau (ZDA) info@zdaarau.ch [copyright holder]