Read in RDB table

Fetches a table from the RDB DBMS's public schema and returns it as a tibble.

Usage

read_tbl(
  tbl_name = tbl_metadata$name,
  connection = connect(),
  disconnect = TRUE
)

Arguments

tbl_name

Table name. One of

"actors"

"options"

"legal_instruments"

"legal_norms"

"referendum_types"

"referendum_clusters"

"referendums"

"referendum_types_legal_norms"

"referendum_types_referendums"

"topics_referendums"

"referendum_titles"

"referendum_questions"

"referendum_positions"

"referendum_votes"

"referendum_sub_votes"

"administrative_units"

"supranational_entities"

"countries"

"subnational_entities"

"municipalities"

"languages"

"topics" .

connection

Database connection to a PostgreSQL DBMS as returned by connect(). An S4 object that inherits from DBI::DBIConnection.

disconnect

Whether or not to terminate the database connection when finished.

Value

A tibble.

See also

Examples

rdb::read_tbl("countries")
#> # A tibble: 281 × 7
#>    code  name                 name_long                       created_by updated_by created_at          updated_at         
#>    <chr> <chr>                <chr>                           <chr>      <chr>      <dttm>              <dttm>             
#>  1 GF    French Guiana        French Guiana                   rdb_admin  rdb_admin  2024-08-19 15:25:09 2024-08-19 15:32:36
#>  2 AD    Andorra              Principality of Andorra         rdb_admin  rdb_admin  2024-08-19 15:25:09 2024-08-19 15:32:36
#>  3 AE    United Arab Emirates United Arab Emirates            rdb_admin  rdb_admin  2024-08-19 15:25:09 2024-08-19 15:32:36
#>  4 AF    Afghanistan          Islamic Republic of Afghanistan rdb_admin  rdb_admin  2024-08-19 15:25:09 2024-08-19 15:32:36
#>  5 AG    Antigua and Barbuda  Antigua and Barbuda             rdb_admin  rdb_admin  2024-08-19 15:25:09 2024-08-19 15:32:36
#>  6 AI    Anguilla             Anguilla                        rdb_admin  rdb_admin  2024-08-19 15:25:09 2024-08-19 15:32:36
#>  7 AL    Albania              Republic of Albania             rdb_admin  rdb_admin  2024-08-19 15:25:09 2024-08-19 15:32:36
#>  8 AM    Armenia              Republic of Armenia             rdb_admin  rdb_admin  2024-08-19 15:25:09 2024-08-19 15:32:36
#>  9 AO    Angola               Republic of Angola              rdb_admin  rdb_admin  2024-08-19 15:25:09 2024-08-19 15:32:36
#> 10 AQ    Antarctica           Antarctica                      rdb_admin  rdb_admin  2024-08-19 15:25:09 2024-08-19 15:32:36
#> # ℹ 271 more rows