Get referendum data

Downloads the referendum data from the Referendum Database (RDB). See the codebook for a detailed description of all variables.

Usage

rfrnds(
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date_min = NULL,
  date_max = NULL,
  is_draft = FALSE,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  date_time_last_edited_min = NULL,
  date_time_last_edited_max = NULL,
  query_filter = NULL,
  incl_archive = FALSE,
  tidy = TRUE,
  use_cache = TRUE,
  max_cache_age = "1 week",
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg),
  quiet = FALSE
)

Arguments

country_code

The country_code(s) to be included. A character vector.

subnational_entity_name

The subnational_entity_name(s) to be included. A character vector.

municipality

The municipality(s) to be included. A character vector.

level

The level(s) to be included. A character vector.

type

The type(s) to be included. A character vector.

date_min

The minimum date to be included. A date or something coercible to.

date_max

The maximum date to be included. A date or something coercible to.

is_draft

TRUE means to include only referendum entries with draft status, FALSE to include only normal entries. Set to NULL in order to include both draft and normal entries.

date_time_created_min

The minimum date_time_created to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_created_max

The maximum date_time_created to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_last_edited_min

The minimum date_time_last_edited to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

date_time_last_edited_max

The maximum date_time_last_edited to be included. A datetime, or something coercible to (like "2006-01-02" or "2006-01-02T15:04:05Z"; assumed to be in UTC if no timezone is given).

query_filter

A valid MongoDB JSON query filter document which allows for maximum control over what data is included. This takes precedence over all of the above listed parameters, i.e. if query_filter is provided, the parameters country_code, subnational_entity_name, municipality, level, type, date_min, date_max, is_draft, date_time_created_min, date_time_created_max, date_time_last_edited_min and date_time_last_edited_max are ignored.

incl_archive

Whether or not to include an archive column containing data from an earlier, obsolete state of the Referendum Database (RDB).

tidy

Whether or not to tidy the referendum data, i.e. apply various data cleansing tasks and add additional variables. If FALSE, the raw MongoDB referendum data will only be modified just enough to be able to return it as a tibble. Note that untidy data doesn't conform to the codebook (i.a. different variable names).

use_cache

Return cached results if possible. If FALSE, results are always newly fetched regardless of max_cache_age.

max_cache_age

The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Use Inf to disable cache expiry. Only relevant if use_cache = TRUE.

use_testing_server

Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.

quiet

Whether or not to suppress printing status output from internal processing.

Value

A tibble.

See also

Examples

if (FALSE) {
# get all referendums (excl. drafts)
rdb::rfrnds()}

# get only referendums in Austria and Australia on subnational level
rdb::rfrnds(country_code = c("AT", "AU"),
            level = "subnational")
#> # A tibble: 31 × 69
#>    id    id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date       is_fo…⁶ title…⁷ title…⁸ title…⁹ quest…˟ quest…˟ commi…˟ result subte…˟ subte…˟ elect…˟
#>    <chr> <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>     <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <fct>    <dbl>   <dbl>   <int>
#>  1 5bbb… 12707   NA      AU      Austra… Northe… NA      subn… 2001-10-24 FALSE   Immedi… NA      NA      NA      NA      NA      yes         NA      NA    1140
#>  2 5bbb… 1219    NA      AU      Austra… Christ… NA      subn… 1999-11-06 FALSE   Enhanc… NA      NA      NA      NA      NA      yes         NA      NA      NA
#>  3 5bbb… 12706   NA      AU      Austra… Northe… NA      subn… 1999-05-12 FALSE   NA      NA      Consti… NA      NA      NA      no          NA      NA    1110
#>  4 5bbb… 1710    NA      AU      Austra… Northe… NA      subn… 1998-08-26 FALSE   Title … NA      NA      NA      NA      NA      no          NA      NA    1114
#>  5 5bbb… 12705   NA      AU      Austra… Northe… NA      subn… 1998-05-14 FALSE   Nomina… NA      NA      NA      NA      NA      no          NA      NA      NA
#>  6 5bbb… 12704   NA      AU      Austra… Northe… NA      subn… 1996-10-17 FALSE   Plans … NA      NA      NA      NA      NA      no          NA      NA      NA
#>  7 5bbb… 12703   NA      AU      Austra… Northe… NA      subn… 1995-03-15 FALSE   Quaran… NA      NA      NA      NA      NA      no          NA      NA      NA
#>  8 5bbb… 872     NA      AU      Austra… Christ… NA      subn… 1994-05-07 FALSE   Nation… NA      NA      NA      NA      NA      no          NA      NA      NA
#>  9 5bbb… 871     NA      AU      Austra… Christ… NA      subn… 1994-05-07 FALSE   Nation… NA      NA      NA      NA      NA      yes         NA      NA      NA
#> 10 5bbb… 12694   NA      AU      Austra… Northe… NA      subn… 1991-01-02 FALSE   NA      NA      Consti… NA      NA      NA      no          NA      NA    1051
#> # … with 21 more rows, 49 more variables: electorate_abroad <int>, votes_yes <int>, votes_no <int>, votes_empty <int>, votes_invalid <int>,
#> #   votes_per_subterritory <list>, lower_house_yes <int>, lower_house_no <int>, lower_house_abstentions <int>, upper_house_yes <int>, upper_house_no <int>,
#> #   upper_house_abstentions <int>, position_government <fct>, topics_tier_1 <list>, topics_tier_2 <list>, topics_tier_3 <list>, remarks <chr>, files <list>,
#> #   url_sudd <chr>, url_swissvotes <chr>, sources <chr>, is_draft <lgl>, date_time_created <dttm>, date_time_last_edited <dttm>, type <fct>,
#> #   inst_legal_basis_type <ord>, inst_has_urgent_legal_basis <lgl>, inst_is_binding <lgl>, inst_is_counter_proposal <lgl>, inst_is_assembly <lgl>,
#> #   inst_trigger_type <fct>, inst_trigger_actor <fct>, inst_trigger_actor_level <ord>, inst_trigger_threshold <chr>, inst_trigger_time_limit <chr>,
#> #   inst_object_type <fct>, inst_object_author <fct>, inst_object_legal_level <fct>, inst_object_revision_extent <fct>, inst_object_revision_modes <list>, …

# get referendums in 2020
rdb::rfrnds(date_min = "2020-01-01",
            date_max = "2020-12-31")
#> # A tibble: 103 × 69
#>    id    id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date       is_fo…⁶ title…⁷ title…⁸ title…⁹ quest…˟ quest…˟ commi…˟ result subte…˟ subte…˟ elect…˟
#>    <chr> <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>     <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <fct>    <dbl>   <dbl>   <int>
#>  1 6010… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12 FALSE   Reject… Ablehn… NA      NA      NA      NA      yes         NA      NA  2.07e7
#>  2 6010… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12 FALSE   Initia… Einlei… NA      NA      NA      NA      yes         NA      NA  2.07e7
#>  3 6010… NA      NA      VE      Venezu… NA      NA      nati… 2020-12-12 FALSE   Free e… Freie … NA      NA      NA      NA      yes         NA      NA  2.07e7
#>  4 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Electi… Wahlbe… NA      NA      NA      NA      yes         NA      NA  2.48e6
#>  5 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Electi… Wahlte… NA      NA      NA      NA      yes         NA      NA  2.48e6
#>  6 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Reduct… Kürzun… NA      NA      NA      NA      yes         NA      NA  2.48e6
#>  7 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Reduct… Kürzun… NA      NA      NA      NA      yes         NA      NA  2.48e6
#>  8 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Reduct… Kürzun… NA      NA      NA      NA      yes         NA      NA  2.48e6
#>  9 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Reduct… Kürzun… NA      NA      NA      NA      yes         NA      NA  2.48e6
#> 10 6010… NA      NA      LR      Liberia NA      NA      nati… 2020-12-08 FALSE   Reduct… Kürzun… NA      NA      NA      NA      yes         NA      NA  2.48e6
#> # … with 93 more rows, 49 more variables: electorate_abroad <int>, votes_yes <int>, votes_no <int>, votes_empty <int>, votes_invalid <int>,
#> #   votes_per_subterritory <list>, lower_house_yes <int>, lower_house_no <int>, lower_house_abstentions <int>, upper_house_yes <int>, upper_house_no <int>,
#> #   upper_house_abstentions <int>, position_government <fct>, topics_tier_1 <list>, topics_tier_2 <list>, topics_tier_3 <list>, remarks <chr>, files <list>,
#> #   url_sudd <chr>, url_swissvotes <chr>, sources <chr>, is_draft <lgl>, date_time_created <dttm>, date_time_last_edited <dttm>, type <fct>,
#> #   inst_legal_basis_type <ord>, inst_has_urgent_legal_basis <lgl>, inst_is_binding <lgl>, inst_is_counter_proposal <lgl>, inst_is_assembly <lgl>,
#> #   inst_trigger_type <fct>, inst_trigger_actor <fct>, inst_trigger_actor_level <ord>, inst_trigger_threshold <chr>, inst_trigger_time_limit <chr>,
#> #   inst_object_type <fct>, inst_object_author <fct>, inst_object_legal_level <fct>, inst_object_revision_extent <fct>, inst_object_revision_modes <list>, …

# get referendums added to the database during the last 30 days
rdb::rfrnds(date_time_created_min = clock::date_today(zone = "UTC") |> clock::add_days(-30L),
            date_time_created_max = clock::date_today(zone = "UTC"))
#> data frame with 0 columns and 0 rows

# provide custom `query_filter` for more complex queries like regex matches
# cf. https://docs.mongodb.com/manual/reference/operator/query/regex/
rdb::rfrnds(query_filter = '{"country_code":{"$regex":"A."}}')
#> # A tibble: 243 × 69
#>    id    id_of…¹ id_sudd count…² count…³ subna…⁴ munic…⁵ level date       is_fo…⁶ title…⁷ title…⁸ title…⁹ quest…˟ quest…˟ commi…˟ result subte…˟ subte…˟ elect…˟
#>    <chr> <chr>   <chr>   <fct>   <fct>   <chr>   <chr>   <ord> <date>     <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <fct>    <dbl>   <dbl>   <int>
#>  1 63bd… NA      as1120… AS      Americ… NA      NA      nati… 2022-11-08 FALSE   Renami… Umbene… NA      NA      NA      NA      yes         NA      NA   14313
#>  2 63bd… NA      as1020… AS      Americ… NA      NA      nati… 2022-11-08 FALSE   Renami… Umbene… NA      NA      NA      NA      yes         NA      NA   14313
#>  3 63bd… NA      as0920… AS      Americ… NA      NA      nati… 2022-11-08 FALSE   Renami… Umbene… NA      NA      NA      NA      yes         NA      NA   14313
#>  4 63bd… NA      as0820… AS      Americ… NA      NA      nati… 2022-11-08 FALSE   Remova… Absetz… NA      NA      NA      NA      yes         NA      NA   14313
#>  5 63bd… NA      as0720… AS      Americ… NA      NA      nati… 2022-11-08 FALSE   Enlarg… Vergrö… NA      NA      NA      NA      no          NA      NA   14313
#>  6 63bd… NA      as0620… AS      Americ… NA      NA      nati… 2022-11-08 FALSE   Voting… Stimmr… NA      NA      NA      NA      yes         NA      NA   14313
#>  7 63bd… NA      as0520… AS      Americ… NA      NA      nati… 2022-11-08 FALSE   Annexa… Anschl… NA      NA      NA      NA      no          NA      NA   14313
#>  8 63bd… NA      as0420… AS      Americ… NA      NA      nati… 2022-11-08 FALSE   Enlarg… Vergrö… NA      NA      NA      NA      no          NA      NA   14313
#>  9 63bd… NA      as0320… AS      Americ… NA      NA      nati… 2022-11-08 FALSE   No obj… Keine … NA      NA      NA      NA      no          NA      NA    6800
#> 10 63bd… NA      as0220… AS      Americ… NA      NA      nati… 2022-11-08 FALSE   No obj… Keine … NA      NA      NA      NA      no          NA      NA    6849
#> # … with 233 more rows, 49 more variables: electorate_abroad <int>, votes_yes <int>, votes_no <int>, votes_empty <int>, votes_invalid <int>,
#> #   votes_per_subterritory <list>, lower_house_yes <int>, lower_house_no <int>, lower_house_abstentions <int>, upper_house_yes <int>, upper_house_no <int>,
#> #   upper_house_abstentions <int>, position_government <fct>, topics_tier_1 <list>, topics_tier_2 <list>, topics_tier_3 <list>, remarks <chr>, files <list>,
#> #   url_sudd <chr>, url_swissvotes <chr>, sources <chr>, is_draft <lgl>, date_time_created <dttm>, date_time_last_edited <dttm>, type <fct>,
#> #   inst_legal_basis_type <ord>, inst_has_urgent_legal_basis <lgl>, inst_is_binding <lgl>, inst_is_counter_proposal <lgl>, inst_is_assembly <lgl>,
#> #   inst_trigger_type <fct>, inst_trigger_actor <fct>, inst_trigger_actor_level <ord>, inst_trigger_threshold <chr>, inst_trigger_time_limit <chr>,
#> #   inst_object_type <fct>, inst_object_author <fct>, inst_object_legal_level <fct>, inst_object_revision_extent <fct>, inst_object_revision_modes <list>, …