Get old referendum data

Downloads the latest backup of the old referendum data from the Referendum Database (RDB). See the codebook for a detailed description of all variables.

Usage

rfrnds_old(
  is_draft = FALSE,
  incl_archive = FALSE,
  use_cache = TRUE,
  max_cache_age = "1 week",
  quiet = FALSE
)

Arguments

incl_archive

Whether or not to include an archive column containing data from an earlier, obsolete state of the Referendum Database (RDB).

use_cache

Whether or not to return cached results if possible. If FALSE, results are always newly fetched regardless of max_cache_age.

max_cache_age

Duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Use Inf to disable cache expiry. Only relevant if use_cache = TRUE.

quiet

Whether or not to suppress printing status output from internal processing.

Value

A tibble.

See also

Examples

# get all referendums (excl. drafts)
rdb::rfrnds_old()
#> Error: 'rfrnds_old' is not an exported object from 'namespace:rdb'