Get referendum data from sudd.ch

Downloads referendum data from sudd.ch.

Usage

sudd_rfrnds(ids_sudd, use_cache = TRUE, max_cache_age = "1 week")

Arguments

ids_sudd

The referendum identifiers assigned by sudd.ch. Either as a character vector or a data frame containing a column id_sudd. NAs are ignored.

use_cache

Return cached results if possible. If FALSE, results are always newly fetched regardless of max_cache_age.

max_cache_age

The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Use Inf to disable cache expiry. Only relevant if use_cache = TRUE.

Value

A tibble. The column names are aligned with those of rfrnds() as closely as possible.

About sudd.ch

sudd stands for Suchmaschine für direkte Demokratie (German) and is operated by Beat Müller. Its database content is licensed under Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0).

See also

Examples

rdb::rfrnd(id = "5bbc045192a21351232e596f")$id_sudd |> rdb::sudd_rfrnds()
#> # A tibble: 1 × 30
#>   id_sudd  countr…¹ count…² is_fo…³ subna…⁴ terri…⁵ terri…⁶ date        year month   day title…⁷ quest…⁸ resul…⁹ resul…˟ adopt…˟ elect…˟ polli…˟ votes…˟ votes…˟
#>   <chr>    <chr>    <chr>   <lgl>   <chr>   <chr>   <chr>   <date>     <int> <int> <int> <chr>   <chr>   <chr>   <chr>   <list>    <int>   <int>   <int>   <int>
#> 1 li012015 LI       Liecht… FALSE   NA      Liecht… unabhä… 2015-12-13  2015    12    13 Kranke… Entsch… angeno… Enderg… <chr>     19649   13059   12832      14
#> # … with 10 more variables: votes_void <int>, votes_invalid <int>, votes_valid <int>, votes_yes <int>, votes_no <int>, votes_option_1 <int>, files <list>,
#> #   remarks <list>, sources <list>, date_last_edited <date>, and abbreviated variable names ¹​country_code, ²​country_name, ³​is_former_country,
#> #   ⁴​subnational_entity_code, ⁵​territory_name_de, ⁶​territory_type_de, ⁷​title_de, ⁸​question_type_de, ⁹​result_de, ˟​result_status_de, ˟​adoption_requirements_de,
#> #   ˟​electorate_total, ˟​polling_cards, ˟​votes_total, ˟​votes_empty

rdb::rfrnds(country_code = "AT") |> rdb::sudd_rfrnds()
#> # A tibble: 5 × 29
#>   id_sudd  countr…¹ count…² is_fo…³ subna…⁴ terri…⁵ terri…⁶ date        year month   day title…⁷ quest…⁸ resul…⁹ resul…˟ adopt…˟ elect…˟ votes…˟ votes…˟ votes…˟
#>   <chr>    <chr>    <chr>   <lgl>   <chr>   <chr>   <chr>   <date>     <int> <int> <int> <chr>   <chr>   <chr>   <chr>   <list>    <int>   <int>   <int>   <int>
#> 1 at012013 AT       Austria FALSE   NA      Österr… unabhä… 2013-01-20  2013     1    20 Berufs… 2 Opti… Allgem… Enderg… <chr>   6378628 3344620   82226 3262394
#> 2 at012013 AT       Austria FALSE   NA      Österr… unabhä… 2013-01-20  2013     1    20 Berufs… 2 Opti… Allgem… Enderg… <chr>   6378628 3344620   82226 3262394
#> 3 at011994 AT       Austria FALSE   NA      Österr… unabhä… 1994-06-12  1994     6    12 Beitri… Entsch… angeno… Enderg… <chr>   5790578 4768401   43570 4724831
#> 4 at011978 AT       Austria FALSE   NA      Österr… unabhä… 1978-11-05  1978    11     5 Bau de… Entsch… verwor… Enderg… <chr>   5083673 3259118   75971 3182697
#> 5 at011938 AT       Austria FALSE   NA      Österr… unabhä… 1938-04-10  1938     4    10 Anschl… Entsch… angeno… Enderg… <chr>   4484617 4471618    5777 4465841
#> # … with 9 more variables: votes_yes <int>, votes_no <int>, votes_option_1 <int>, votes_option_2 <int>, votes_option_3 <int>, files <list>, remarks <list>,
#> #   sources <list>, date_last_edited <date>, and abbreviated variable names ¹​country_code, ²​country_name, ³​is_former_country, ⁴​subnational_entity_code,
#> #   ⁵​territory_name_de, ⁶​territory_type_de, ⁷​title_de, ⁸​question_type_de, ⁹​result_de, ˟​result_status_de, ˟​adoption_requirements_de, ˟​electorate_total,
#> #   ˟​votes_total, ˟​votes_invalid, ˟​votes_valid