Unnest multi-value variable

Unnests a multi-value variable of type list to long format. Multi-value variables can contain more than one value per observation and thus break with the tidy-data convention. This function allows to conveniently expand data to contain a single var value per observation only, thereby increasing the number of observations (i.e. rows).

Usage

unnest_var(data, var)

Arguments

data

RDB referendum data as returned by rfrnds(). A data frame that at minimum contains the column specified in var.

var

data column to unnest. One of the multi-value variables:

topics_tier_1

topics_tier_2

topics_tier_3

inst_object_revision_modes

inst_topics_only

inst_topics_excluded

Tidy selections are supported.

Value

A tibble.

See also

Examples

rdb::rfrnds(country_code = "AT",
            quiet = TRUE) |>
  rdb::unnest_var(topics_tier_2)
#> # A tibble: 16 × 69
#>    id            id_official id_sudd country_code country_name subnational_entity_n…¹ municipality level date       is_former_country title_en title_de title_fr
#>    <chr>         <chr>       <chr>   <fct>        <fct>        <chr>                  <chr>        <ord> <date>     <lgl>             <chr>    <chr>    <chr>   
#>  1 5bbc025f92a2… NA          at0120… AT           Austria      NA                     NA           nati… 2013-01-20 FALSE             Future … Berufsh… NA      
#>  2 5bbc025f92a2… NA          at0120… AT           Austria      NA                     NA           nati… 2013-01-20 FALSE             Future … Berufsh… NA      
#>  3 5bbbf50592a2… NA          at0119… AT           Austria      NA                     NA           nati… 1994-06-12 FALSE             Accessi… NA       NA      
#>  4 5bbbe98892a2… NA          at0119… AT           Austria      NA                     NA           nati… 1978-11-05 FALSE             Constru… Bau des… NA      
#>  5 5bbbe2a492a2… NA          at0119… AT           Austria      NA                     NA           nati… 1938-04-10 FALSE             Accessi… Anschlu… NA      
#>  6 5bbbe2a492a2… NA          at0119… AT           Austria      NA                     NA           nati… 1938-04-10 FALSE             Accessi… Anschlu… NA      
#>  7 5bbbe2a492a2… NA          at0119… AT           Austria      NA                     NA           nati… 1938-04-10 FALSE             Accessi… Anschlu… NA      
#>  8 5bbbe28792a2… 123         NA      AT           Austria      Salzburg               NA           subn… 1921-05-29 FALSE             Accessi… NA       NA      
#>  9 5bbbe28792a2… 123         NA      AT           Austria      Salzburg               NA           subn… 1921-05-29 FALSE             Accessi… NA       NA      
#> 10 5bbbe28792a2… 122         NA      AT           Austria      Tyrol                  NA           subn… 1921-04-24 FALSE             Accessi… NA       NA      
#> 11 5bbbe28792a2… 122         NA      AT           Austria      Tyrol                  NA           subn… 1921-04-24 FALSE             Accessi… NA       NA      
#> 12 5bbbe28592a2… 5644        NA      AT           Austria      Carinthia              NA           subn… 1920-10-10 FALSE             Nationa… NA       NA      
#> 13 5bbbe28592a2… 5643        NA      AT           Austria      Carinthia              NA           subn… 1920-10-10 FALSE             Nationa… NA       NA      
#> 14 5bbbe28592a2… 114         NA      AT           Austria      Klagenfurt Becken Zon… NA           subn… 1920-10-10 FALSE             Adhesio… NA       NA      
#> 15 5bbbe28292a2… 101         NA      AT           Austria      Vorarlberg             NA           subn… 1919-05-11 FALSE             Accessi… NA       NA      
#> 16 5bbbe28292a2… 101         NA      AT           Austria      Vorarlberg             NA           subn… 1919-05-11 FALSE             Accessi… NA       NA      
#> # ℹ abbreviated name: ¹​subnational_entity_name
#> # ℹ 56 more variables: question <chr>, question_en <chr>, committee_name <chr>, result <fct>, subterritories_yes <dbl>, subterritories_no <dbl>,
#> #   electorate_total <int>, electorate_abroad <int>, votes_yes <int>, votes_no <int>, votes_empty <int>, votes_invalid <int>, votes_per_subterritory <list>,
#> #   lower_house_yes <int>, lower_house_no <int>, lower_house_abstentions <int>, upper_house_yes <int>, upper_house_no <int>, upper_house_abstentions <int>,
#> #   position_government <fct>, topics_tier_1 <list>, topic_tier_2 <fct>, topics_tier_3 <list>, remarks <chr>, files <list>, url_sudd <chr>,
#> #   url_swissvotes <chr>, sources <chr>, is_draft <lgl>, date_time_created <dttm>, date_time_last_edited <dttm>, type <fct>, inst_legal_basis_type <ord>,
#> #   inst_has_urgent_legal_basis <lgl>, inst_is_binding <lgl>, inst_is_counter_proposal <lgl>, inst_is_assembly <lgl>, inst_trigger_type <fct>, …